+++ src/classes/babies/CursedPillowBaby.ts
@@ -1,5 +1,15 @@
-import { DamageFlagZero, ModCallback } from "isaac-typescript-definitions";
-import { Callback, isMissedTear } from "isaacscript-common";
+import {
+  DamageFlagZero,
+  EntityType,
+  GridEntityType,
+  ModCallback,
+} from "isaac-typescript-definitions";
+import {
+  Callback,
+  CallbackCustom,
+  ModCallbackCustom,
+  isMissedTear,
+} from "isaacscript-common";
 import {
   getBabyPlayerFromEntity,
   isValidForMissedTearsEffect,
@@ -44,10 +54,27 @@ export class CursedPillowBaby extends Baby {
     const num = this.getAttribute("num");
 
     v.run.numTearMisses++;
+    Isaac.DebugString(`GETTING HERE 2 - ${v.run.numTearMisses}`);
     if (v.run.numTearMisses === num) {
       v.run.numTearMisses = 0;
       player.TakeDamage(1, DamageFlagZero, EntityRef(player), 0);
     }
+  }
+
+  // 42
+  @Callback(ModCallback.PRE_TEAR_COLLISION)
+  preTearCollision(
+    tear: EntityTear,
+    collider: Entity,
+    _low: boolean,
+  ): boolean | undefined {
+    if (collider.Type !== EntityType.FIREPLACE) {
+      return undefined;
+    }
+
+    const ptrHash = GetPtrHash(tear);
+    v.room.tearPtrHashes.delete(ptrHash);
+    return undefined;
   }
 
   // 61
@@ -55,5 +82,16 @@ export class CursedPillowBaby extends Baby {
   postFireTear(tear: EntityTear): void {
     const ptrHash = GetPtrHash(tear);
     v.room.tearPtrHashes.add(ptrHash);
+  }
+
+  @CallbackCustom(
+    ModCallbackCustom.POST_GRID_ENTITY_COLLISION,
+    GridEntityType.POOP,
+    undefined,
+    EntityType.TEAR,
+  )
+  postGridEntityCollisionPoop(_gridEntity: GridEntity, entity: Entity): void {
+    const ptrHash = GetPtrHash(entity);
+    v.room.tearPtrHashes.delete(ptrHash);
   }
 }